cloud-function/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
decorator==4.4.2          # via gcsfs
fsspec==0.7.4             # via gcsfs
gcsfs==0.6.2              # via google-datacatalog-rdbms-connector
google-api-core[grpc]==1.21.0  # via google-cloud-core, google-cloud-datacatalog, google-cloud-logging, google-cloud-monitoring, google-cloud-secret-manager
google-auth-oauthlib==0.4.1  # via gcsfs
google-auth==1.18.0       # via gcsfs, google-api-core, google-auth-oauthlib
google-cloud-core==1.3.0  # via google-cloud-logging
google-cloud-datacatalog==1.0.0  # via google-datacatalog-connectors-commons
google-cloud-logging==1.15.0  # via -r requirements.in
google-cloud-monitoring==1.0.0  # via google-datacatalog-connectors-commons
google-cloud-secret-manager==0.1.1  # via -r requirements.in
google-datacatalog-connectors-commons==0.5.0  # via google-datacatalog-rdbms-connector
google-datacatalog-postgresql-connector==0.6.0  # via -r requirements.in
google-datacatalog-rdbms-connector==0.6.0  # via google-datacatalog-postgresql-connector
googleapis-common-protos[grpc]==1.52.0  # via google-api-core, grpc-google-iam-v1
grpc-google-iam-v1==0.12.3  # via google-cloud-datacatalog, google-cloud-secret-manager
grpcio==1.30.0            # via google-api-core, googleapis-common-protos, grpc-google-iam-v1
idna==2.10                # via requests
numpy==1.19.1             # via pandas
oauthlib==3.1.0           # via requests-oauthlib
pandas==0.24.2            # via google-datacatalog-rdbms-connector
protobuf==3.12.2          # via google-api-core, googleapis-common-protos
psycopg2-binary==2.8.5    # via google-datacatalog-postgresql-connector
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-dateutil==2.8.1    # via google-datacatalog-connectors-commons, pandas
pytz==2020.1              # via google-api-core, pandas
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.24.0          # via gcsfs, google-api-core, requests-oauthlib
rsa==4.6                  # via google-auth
six==1.14.0               # via google-api-core, google-auth, grpcio, protobuf, python-dateutil
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
